Other_Agents/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
google-cloud-aiplatform==1.75.0
google-cloud-storage==2.18.2
google-generativeai==0.8.3
Pillow==10.4.0
